Initial commit
This is a torch plugin for Jolla settings.
It ships with a switch that turns on
the flashlight.

Torch uses torchd, an userspace daemon that
is invoked via DBus. torchd controls the
torch using GStreamer.

The settings plugin uses DBus to communicate
to the daemon, turning the flashlight when
the switch is pressed.

When the flashlight is off, the daemon shouldn't
use any resource, so other camera-related
applications should work.